fix: mudei o DOUBLE para DOUBLE PRECISION, porque e compativel com postgreSQL

+++ project/src/main/resources/db/migration/V1__create_initial_schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE pets (
     species     VARCHAR(20)  NOT NULL,
     breed       VARCHAR(100),
     birth_date  DATE,
-    weight      DOUBLE,
+    weight      DOUBLE PRECISION,
     tutor_name  VARCHAR(100) NOT NULL,
     tutor_phone VARCHAR(20),
     active      BOOLEAN      NOT NULL DEFAULT TRUE,
